requirements.txt
uqlm
streamlit
langchain-google-genai
pandas
plotly
asyncio
python-dotenv
nest_asyncio